chore: post-0.7.0 internal cleanup — Windows clippy casts + dead-code triage (#129, #125)
Non-functional cleanup of debt from the 0.7.0 API-narrowing refactor (#67). No behavior change; data path untouched.

- #129: drop redundant `as i32` casts on windows-sys socket constants (caught only by windows-target clippy).
- #125: remove unused DEFAULT_TIMESTAMP_FORMAT; gate test-only print_summary + snapshot_omit to #[cfg(test)]; keep documented parity allows.

Closes #129
Closes #125

+++ riperf3/src/net.rs
@@ -389,9 +389,9 @@ pub fn configure_tcp_stream_full(
             unsafe {
                 windows_sys::Win32::Networking::WinSock::setsockopt(
                     raw as usize,
-                    windows_sys::Win32::Networking::WinSock::IPPROTO_TCP as i32,
-                    windows_sys::Win32::Networking::WinSock::TCP_MAXSEG as i32,
-                    &(mss_val as i32) as *const i32 as *const u8,
+                    windows_sys::Win32::Networking::WinSock::IPPROTO_TCP,
+                    windows_sys::Win32::Networking::WinSock::TCP_MAXSEG,
+                    &mss_val as *const i32 as *const u8,
                     std::mem::size_of::<i32>() as i32,
                 );
             }
@@ -569,8 +569,8 @@ pub fn set_dont_fragment(fd: &impl std::os::windows::io::AsSocket) -> Result<()>
     let ret = unsafe {
         windows_sys::Win32::Networking::WinSock::setsockopt(
             raw as usize,
-            windows_sys::Win32::Networking::WinSock::IPPROTO_IP as i32,
-            windows_sys::Win32::Networking::WinSock::IP_DONTFRAGMENT as i32,
+            windows_sys::Win32::Networking::WinSock::IPPROTO_IP,
+            windows_sys::Win32::Networking::WinSock::IP_DONTFRAGMENT,
             &val as *const i32 as *const u8,
             std::mem::size_of::<i32>() as i32,
         )
